+++ tests/conv_on_coords.py
@@ -1,0 +1,105 @@
+# Copyright (c) Chris Choy (chrischoy@ai.stanford.edu).
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy of
+# this software and associated documentation files (the "Software"), to deal in
+# the Software without restriction, including without limitation the rights to
+# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+# of the Software, and to permit persons to whom the Software is furnished to do
+# so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+# Please cite "4D Spatio-Temporal ConvNets: Minkowski Convolutional Neural
+# Networks", CVPR'19 (https://arxiv.org/abs/1904.08755) if you use any part
+# of the code.
+import torch
+import unittest
+
+from MinkowskiEngine import SparseTensor, MinkowskiConvolutionOnCoords, \
+    MinkowskiConvolutionTransposeOnCoords
+
+from tests.common import data_loader
+
+
+class TestConvolution(unittest.TestCase):
+
+    def test(self):
+        print(f"{self.__class__.__name__}: test")
+        in_channels, out_channels, D = 2, 3, 2
+        coords, feats, labels = data_loader(in_channels, batch_size=2)
+
+        # Create random coordinates with tensor stride == 2
+        out_coords = torch.rand(10, 3)
+        out_coords[:, :2] *= 5  # random coords
+        out_coords[:, 2] *= 2  # random batch index
+        out_coords = out_coords.floor().int()
+        out_coords[:, :2] *= 2  # make the tensor stride 2
+
+        feats = feats.double()
+        feats.requires_grad_()
+        input = SparseTensor(feats, coords=coords)
+        print(input.mapping)
+        cm = input.coords_man
+        print(cm.get_coords_key(1))
+
+        conv = MinkowskiConvolutionOnCoords(
+            in_channels,
+            out_channels,
+            kernel_size=3,
+            stride=1,
+            has_bias=False,
+            dimension=D).double()
+
+        print('Initial input: ', input)
+        print('Specified output coords: ', out_coords)
+        output = conv(input, out_coords, tensor_stride=2)
+        print('Conv output: ', output)
+
+        output.F.sum().backward()
+        print(input.F.grad)
+
+    def test_tr(self):
+        print(f"{self.__class__.__name__}: test")
+        in_channels, out_channels, D = 2, 3, 2
+        coords, feats, labels = data_loader(in_channels, batch_size=2)
+        # tensor stride must be at least 2 for convolution transpose with stride 2
+        coords[:, :2] *= 2
+        out_coords = torch.rand(10, 3)
+        out_coords[:, :2] *= 10  # random coords
+        out_coords[:, 2] *= 2  # random batch index
+        out_coords = out_coords.floor().int()
+
+        feats = feats.double()
+        feats.requires_grad_()
+        input = SparseTensor(feats, coords=coords, tensor_stride=2)
+        print(input.mapping)
+        cm = input.coords_man
+        print(cm.get_coords_key(2))
+
+        conv_tr = MinkowskiConvolutionTransposeOnCoords(
+            in_channels,
+            out_channels,
+            kernel_size=3,
+            stride=2,
+            has_bias=False,
+            dimension=D).double()
+        print('Initial input: ', input)
+        print('Specified output coords: ', out_coords)
+        output = conv_tr(input, out_coords)
+        print('Conv output: ', output)
+
+        output.F.sum().backward()
+        print(input.F.grad)
+
+
+if __name__ == '__main__':
+    unittest.main()